pyDKB/common: move `read_es_config()` function from stage to library.
Function like this seem to be needed in multiple places, so why not to
have it in the common library.
