+++ Utils/Dataflow/091_datasetsRucio/datasets_processing.py
@@ -37,6 +37,7 @@ try:
     sys.path.append(dkb_dir)
     import pyDKB
     from pyDKB import storages
+    from pyDKB.common.utils import read_es_config
 except Exception, err:
     sys.stderr.write("(ERROR) Failed to import pyDKB library: %s\n" % err)
     sys.exit(1)
@@ -101,34 +102,6 @@ def main(argv):
         stage.stop()
 
     sys.exit(exit_code)
-
-
-def read_es_config(cfg_file):
-    """ Read ES configuration file.
-
-    :param cfg_file: open file descriptor with ES access configuration
-    :type cfg_file: file descriptor
-    """
-    keys = {'ES_HOST': 'host',
-            'ES_PORT': 'port',
-            'ES_USER': 'user',
-            'ES_PASSWORD': '__passwd',
-            'ES_INDEX': 'index'
-            }
-    cfg = {}
-    for line in cfg_file.readlines():
-        if line.strip().startswith('#'):
-            continue
-        line = line.split('#')[0].strip()
-        if '=' not in line:
-            continue
-        key, val = line.split('=')[:2]
-        try:
-            cfg[keys[key]] = val
-        except KeyError:
-            sys.stderr.write("(WARN) Unknown configuration parameter: "
-                             "'%s'.\n" % key)
-    return cfg
 
 
 def init_rucio_client():
